requirements.txt
tqdm~=4.62.3
colorama~=0.4.4
PyNaCl==1.5.0